requirements.txt
pytz
psycopg2
python-dateutil
python-dotenv
IMDbPY
discord.py
praw